+++ src/mjlab/scripts/export_scene.py
@@ -17,6 +17,7 @@ from mjlab.utils.lab_api.string import string_to_callable
 
 ENTITY_ALIASES: dict[str, str] = {
   "g1": "mjlab.asset_zoo.robots:get_g1_robot_cfg",
+  "g1_29dof_mode_15": "mjlab.asset_zoo.robots:get_g1_29dof_mode_15_robot_cfg",
   "go1": "mjlab.asset_zoo.robots:get_go1_robot_cfg",
   "yam": "mjlab.asset_zoo.robots:get_yam_robot_cfg",
 }