+++ src/mjlab/tasks/tracking/config/g1_29dof_mode_15/__init__.py
@@ -1,0 +1,23 @@
+from mjlab.tasks.registry import register_mjlab_task
+from mjlab.tasks.tracking.rl import MotionTrackingOnPolicyRunner
+
+from .env_cfgs import unitree_g1_29dof_mode_15_flat_tracking_env_cfg
+from .rl_cfg import unitree_g1_29dof_mode_15_tracking_ppo_runner_cfg
+
+register_mjlab_task(
+  task_id="Mjlab-Tracking-Flat-Unitree-G1-29DOF-Mode-15",
+  env_cfg=unitree_g1_29dof_mode_15_flat_tracking_env_cfg(),
+  play_env_cfg=unitree_g1_29dof_mode_15_flat_tracking_env_cfg(play=True),
+  rl_cfg=unitree_g1_29dof_mode_15_tracking_ppo_runner_cfg(),
+  runner_cls=MotionTrackingOnPolicyRunner,
+)
+
+register_mjlab_task(
+  task_id="Mjlab-Tracking-Flat-Unitree-G1-29DOF-Mode-15-No-State-Estimation",
+  env_cfg=unitree_g1_29dof_mode_15_flat_tracking_env_cfg(has_state_estimation=False),
+  play_env_cfg=unitree_g1_29dof_mode_15_flat_tracking_env_cfg(
+    has_state_estimation=False, play=True
+  ),
+  rl_cfg=unitree_g1_29dof_mode_15_tracking_ppo_runner_cfg(),
+  runner_cls=MotionTrackingOnPolicyRunner,
+)